typescript

+++ src/routers/team.ts
@@ -1,18 +1,21 @@
 import { Router } from "express";
-import Team from '../models/team.mjs'
-import Category from '../models/category.mjs'
-import auth from '../middleware/auth.mjs'
-import Register from "../models/register.mjs";
-const router = new Router()
+import Team from '../models/team.js'
+import auth from '../middleware/auth.js'
+import Register from "../models/register.js";
+import { MyRequest } from "../interfaces/MyRequest.js";
+const router = Router()
 
 // Create team
-router.post('/teams',auth, async (req, res) => {
+router.post('/teams',auth, async (req: MyRequest, res) => {
     try {
+        if (typeof req.gymOwner === "undefined") {
+            return res.status(400).send()
+        }
         const register = await Register.findOne({ _id: req.body.registerId, gymOwnerId: req.gymOwner._id }).populate('event teams')
         if (!register) {
             return res.status(404).send('Register not found!')
         }
-        if (!register.event.categoriesIds.includes(req.body.categoryId)) {
+        if (register.event && !register.event.categoriesIds.includes(req.body.categoryId)) {
             return res.status(400).send('Categories doesnt match!')
         }
         const team = new Team({
@@ -24,8 +27,8 @@ router.post('/teams',auth, async (req, res) => {
         })
         await team.save()
         res.status(201).send(team)
-    } catch (err) {
-        res.status(400).send(err)
+    } catch (err: any) {
+        res.status(400).send({ error: err.message })
     }
 })
 
@@ -33,47 +36,56 @@ router.post('/teams',auth, async (req, res) => {
 // GET /teams?category=Football
 // GET /teams?limit=10&skip=0
 // GET /teams?sortBy=createdAt:asc
-router.get('/teams', auth, async (req, res) => {
-    const match = {}
-    const sort = {}
-    if (req.query.category) {
-        match.categoryId = await Category.findOne({ name: req.query.category.toLowerCase() })
+router.get('/teams', auth, async (req: MyRequest, res) => {
+    const match: {categoryId?: string} = {}
+    const sort: {sortBy?: number} = {}
+    if (typeof req.query.categoryId === 'string') {
+        match.categoryId = req.query.categoryId
     }
     if (req.query.sortBy) {
-        const [sortBy, order] = req.query.sortBy.split(':')
-        sort[sortBy] = order === 'desc' ? -1 : 1
+        if (typeof req.query.sortBy === 'string') {
+            const [sortBy, order] = req.query.sortBy.split(':')
+            
+            sort.sortBy = order === 'desc' ? -1 : 1
+        }
     }
     try {
+        if (typeof req.gymOwner === "undefined") {
+            return res.status(400).send()
+        }
         await req.gymOwner.populate({
             path: 'teams',
             match,
             options: {
-                limit: parseInt(req.query.limit),
-                skip: parseInt(req.query.skip),
+                limit: parseInt(typeof req.query.limit === 'string' ? req.query.limit : ""),
+                skip: parseInt(typeof req.query.skip === 'string' ? req.query.skip : ""),
                 sort
             }
         }).execPopulate()
         res.send(req.gymOwner.teams)
-    } catch (err) {
-        res.status(400).send(err)
+    } catch (err: any) {
+        res.status(400).send({ error: err.message })
     }
 })
 
 // Read team by ID
-router.get('/teams/:id', auth, async (req, res) => {
+router.get('/teams/:id', auth, async (req: MyRequest, res) => {
     try {
+        if (typeof req.gymOwner === "undefined") {
+            return res.status(400).send('GymOwner not available')
+        }
         const team = await Team.findOne({ _id: req.params.id, gymOwnerId: req.gymOwner._id }).populate('category event')
         if (!team) {
             return res.status(404).send('Team not found!')
         }
         res.send(team)
-    } catch (err) {
-        res.status(500).send(err)
+    } catch (err: any) {
+        res.status(400).send({ error: err.message })
     }
 })
 
 // Update team by ID
-router.patch('/teams/:id', auth, async (req, res) => {
+router.patch('/teams/:id', auth, async (req: MyRequest, res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ['name', 'categoryId', 'image']
     const isValidOp = updates.every((update) => allowedUpdates.includes(update))
@@ -81,32 +93,38 @@ router.patch('/teams/:id', auth, async (req, res) => {
         return res.status(400).send()
     }
     try {
+        if (typeof req.gymOwner === "undefined") {
+            return res.status(400).send('GymOwner not available')
+        }
         const team = await Team.findOne({ _id: req.params.id, gymOwnerId: req.gymOwner._id })
         const register = await Register.findOne({ _id: team.registerId, gymOwnerId: req.gymOwner._id }).populate('event')
-        if (!team) {
+        if (!team || !register) {
             return res.status(404).send()
         }
-        if (req.body.categoryId && !register.event.categoriesIds.includes(req.body.categoryId)) {
+        if (req.body.categoryId && register.event && !register.event.categoriesIds.includes(req.body.categoryId)) {
             return res.status(400).send('Categories doesnt match!')
         }
         updates.forEach((update) => team[update] = req.body[update])
         await team.save()
         res.send(team)
-    } catch (err) {
-        res.status(400).send(err)
+    } catch (err: any) {
+        res.status(400).send({ error: err.message })
     }
 })
 
 // Delete team by ID
-router.delete('/teams/:id', auth, async (req, res) => {
+router.delete('/teams/:id', auth, async (req: MyRequest, res) => {
     try {
+        if (typeof req.gymOwner === "undefined") {
+            return res.status(400).send('GymOwner not available')
+        }
         const team = await Team.findOneAndDelete({ _id: req.params.id, gymOwnerId: req.gymOwner._id })
         if (!team) {
             res.status(404).send()
         }
         res.send()
-    } catch (err) {
-        res.status(500).send(err)
+    } catch (err: any) {
+        res.status(500).send({ error: err.message })
     }
 })
